Добавлен параметр, отключающий Elasticsearch. Основное назначение - тестирование. Потому что, например, модель, которая индексируется в поиске, во время тестирования, которое не рассчитано на ES вываливает огромное количество ошибок.

+++ efsw/common/signals.py
@@ -1,15 +1,18 @@
 from django.dispatch import receiver
 from django.db.models import signals
+from django.conf import settings
 
 from elasticsearch.exceptions import NotFoundError as EsNotFoundError
 
 from efsw.common.search.models import IndexableModel
 from efsw.common.search import elastic
+from efsw.common import default_settings
 
 
 @receiver(signals.post_save)
 def model_saved(sender, instance, created, raw, *args, **kwargs):
-    if isinstance(instance, IndexableModel):
+    if isinstance(instance, IndexableModel)\
+            and not getattr(settings, "EFSW_ELASTIC_DISABLE", default_settings.EFSW_ELASTIC_DISABLE):
         if created:
             elastic.create_document(instance)
         else:
@@ -21,5 +24,6 @@ def model_saved(sender, instance, created, raw, *args, **kwargs):
 
 @receiver(signals.post_delete)
 def model_deleted(sender, instance, *args, **kwargs):
-    if isinstance(instance, IndexableModel):
+    if isinstance(instance, IndexableModel)\
+            and not getattr(settings, "EFSW_ELASTIC_DISABLE", default_settings.EFSW_ELASTIC_DISABLE):
         elastic.delete_document(instance)
